requirements.txt

numpy
scipy
python_dateutil

jupyter
matplotlib
pandas
librosa
bokeh

